requirements.txt
absl-py==2.1.0
asgiref==3.7.2
attrs==23.2.0
autopep8==2.2.0
cffi==1.16.0
contourpy==1.2.1
cycler==0.12.1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
flatbuffers==24.3.25
fonttools==4.53.0
jax==0.4.28
jaxlib==0.4.28
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
mediapipe==0.10.14
ml-dtypes==0.4.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pycodestyle==2.11.1
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sounddevice==0.4.7
sqlparse==0.4.4
threadpoolctl==3.5.0
typing_extensions==4.9.0
